+++ discord/ext/commands/__init__.py
@@ -5,6 +5,7 @@ discord.ext.commands
 An extension module to facilitate creation of bot commands.
 
 :copyright: (c) 2015-2021 Rapptz
+:copyright: (c) 2021-present Kae Bartlett
 :license: MIT, see LICENSE for more details.
 """
 
@@ -17,3 +18,4 @@ from .converter import *
 from .cooldowns import *
 from .cog import *
 from .flags import *
+from .http_interactions import *