+++ discord/interactions.py
@@ -28,6 +28,9 @@ DEALINGS IN THE SOFTWARE.
 from __future__ import annotations
 from typing import Any, Dict, List, Optional, TYPE_CHECKING, Tuple, Union
 import asyncio
+import json
+
+from aiohttp import web
 
 from . import utils
 from .enums import try_enum, InteractionType, InteractionResponseType
@@ -286,7 +289,7 @@ class Interaction:
         return Permissions(self._permissions)
 
     @utils.cached_slot_property('_cs_response')
-    def response(self) -> InteractionResponse:
+    def response(self) -> Union[InteractionResponse, HTTPInteractionResponse]:
         """:class:`InteractionResponse`: Returns an object responsible for handling responding to the interaction.
 
         A response can only be done once. If secondary messages need to be sent, consider using :attr:`followup`
@@ -763,6 +766,315 @@ class InteractionResponse:
             type=InteractionResponseType.message_update.value,
             data=payload,
         )
+
+        self._responded = True
+
+
+class HTTPInteractionResponse(InteractionResponse):
+    """Represents a Discord interaction response.
+
+    This type can be accessed through :attr:`Interaction.response`.
+    """
+
+    __slots__: Tuple[str, ...] = (
+        '_responded',
+        '_parent',
+        '_aiohttp_request',
+        '_aiohttp_response',
+    )
+
+    def __init__(self, aiohttp_response: web.Request, parent: Interaction):
+        super().__init__(parent)
+        self._aiohttp_request: web.Request = aiohttp_response
+        self._aiohttp_response: web.StreamResponse = web.StreamResponse()
+
+    async def defer(self, *, ephemeral: bool = False) -> None:
+        """|coro|
+
+        Defers the interaction response.
+
+        This is typically used when the interaction is acknowledged
+        and a secondary action will be done later.
+
+        Parameters
+        -----------
+        ephemeral: :class:`bool`
+            Indicates whether the deferred message will eventually be ephemeral.
+            This only applies for interactions of type :attr:`InteractionType.application_command`.
+
+        Raises
+        -------
+        HTTPException
+            Deferring the interaction failed.
+        InteractionResponded
+            This interaction has already been responded to before.
+        """
+        if self._responded:
+            raise InteractionResponded(self._parent)
+
+        defer_type: int = 0
+        data: Optional[Dict[str, Any]] = None
+        defer_type = InteractionResponseType.deferred_channel_message.value
+        if ephemeral:
+            data = {'flags': 64}
+
+        if defer_type:
+            payload = {"type": defer_type}
+            if data:
+                payload["data"] = data
+            self._aiohttp_response.headers["Content-Type"] = "application/json"
+            await self._aiohttp_response.prepare(self._aiohttp_request)
+            await self._aiohttp_response.write(json.dumps(payload).encode())
+            await self._aiohttp_response.write_eof()
+            self._responded = True
+
+    async def defer_update(self) -> None:
+        """|coro|
+
+        Defers the interaction response.
+
+        This is typically used when the interaction is acknowledged
+        and a secondary action will be done later.
+
+        This is only usable if the interaction spawned from a component interaction.
+
+        Raises
+        -------
+        HTTPException
+            Deferring the interaction failed.
+        InteractionResponded
+            This interaction has already been responded to before.
+        """
+        if self._responded:
+            raise InteractionResponded(self._parent)
+
+        defer_type: int = 0
+        defer_type = InteractionResponseType.deferred_message_update.value
+
+        if defer_type:
+            payload = {"type": defer_type}
+            self._aiohttp_response.headers["Content-Type"] = "application/json"
+            await self._aiohttp_response.prepare(self._aiohttp_request)
+            await self._aiohttp_response.write(json.dumps(payload).encode())
+            await self._aiohttp_response.write_eof()
+            self._responded = True
+
+    async def pong(self) -> None:
+        """|coro|
+
+        Pongs the ping interaction.
+
+        This should rarely be used.
+
+        Raises
+        -------
+        HTTPException
+            Ponging the interaction failed.
+        InteractionResponded
+            This interaction has already been responded to before.
+        """
+        if self._responded:
+            raise InteractionResponded(self._parent)
+
+        parent = self._parent
+        if parent.type is InteractionType.ping:
+            payload = {"type": InteractionResponseType.pong.value}
+            self._aiohttp_response.headers["Content-Type"] = "application/json"
+            await self._aiohttp_response.prepare(self._aiohttp_request)
+            await self._aiohttp_response.write(json.dumps(payload).encode())
+            await self._aiohttp_response.write_eof()
+            self._responded = True
+
+    async def send_message(
+        self,
+        content: Optional[Any] = None,
+        *,
+        embed: Embed = MISSING,
+        embeds: List[Embed] = MISSING,
+        components: MessageComponents = MISSING,
+        tts: bool = False,
+        ephemeral: bool = False,
+        allowed_mentions: AllowedMentions = None,
+    ) -> None:
+        """|coro|
+
+        Responds to this interaction by sending a message.
+
+        Parameters
+        -----------
+        content: Optional[:class:`str`]
+            The content of the message to send.
+        embeds: List[:class:`Embed`]
+            A list of embeds to send with the content. Maximum of 10. This cannot
+            be mixed with the ``embed`` parameter.
+        embed: :class:`Embed`
+            The rich embed for the content to send. This cannot be mixed with
+            ``embeds`` parameter.
+        tts: :class:`bool`
+            Indicates if the message should be sent using text-to-speech.
+        components: :class:`discord.ui.MessageComponents`
+            The components to send with the messasge.
+        ephemeral: :class:`bool`
+            Indicates if the message should only be visible to the user who started the interaction.
+        allowed_mentions: :class:`AllowedMentions`
+            The allowed mentions that should be sent with the message.
+
+        Raises
+        -------
+        HTTPException
+            Sending the message failed.
+        TypeError
+            You specified both ``embed`` and ``embeds``.
+        ValueError
+            The length of ``embeds`` was invalid.
+        InteractionResponded
+            This interaction has already been responded to before.
+        """
+        if self._responded:
+            raise InteractionResponded(self._parent)
+
+        payload: Dict[str, Any] = {
+            'tts': tts,
+        }
+
+        if embed is not MISSING and embeds is not MISSING:
+            raise TypeError('cannot mix embed and embeds keyword arguments')
+
+        if embed is not MISSING:
+            embeds = [embed]
+
+        if embeds:
+            if len(embeds) > 10:
+                raise ValueError('embeds cannot exceed maximum of 10 elements')
+            payload['embeds'] = [e.to_dict() for e in embeds]
+
+        if content is not None:
+            payload['content'] = str(content)
+
+        if ephemeral:
+            payload['flags'] = 64
+
+        if components is not MISSING:
+            payload['components'] = components.to_dict()
+
+        parent = self._parent
+        parent_state = parent._state
+
+        if allowed_mentions is not None:
+            if parent_state.allowed_mentions is not None:
+                allowed_mentions = parent_state.allowed_mentions.merge(allowed_mentions).to_dict()
+            else:
+                allowed_mentions = allowed_mentions.to_dict()
+        else:
+            allowed_mentions = parent_state.allowed_mentions and parent_state.allowed_mentions.to_dict()
+
+        data = {
+            "type": InteractionResponseType.channel_message.value,
+            "data": payload,
+        }
+        self._aiohttp_response.headers["Content-Type"] = "application/json"
+        await self._aiohttp_response.prepare(self._aiohttp_request)
+        await self._aiohttp_response.write(json.dumps(data).encode())
+        await self._aiohttp_response.write_eof()
+
+        self._responded = True
+
+    async def edit_message(
+        self,
+        *,
+        content: Optional[Any] = MISSING,
+        embed: Optional[Embed] = MISSING,
+        embeds: List[Embed] = MISSING,
+        attachments: List[Attachment] = MISSING,
+        components: Optional[MessageComponents] = MISSING,
+        allowed_mentions: Optional[AllowedMentions] = MISSING,
+    ) -> None:
+        """|coro|
+
+        Responds to this interaction by editing the original message of
+        a component interaction.
+
+        Parameters
+        -----------
+        content: Optional[:class:`str`]
+            The new content to replace the message with. ``None`` removes the content.
+        embeds: List[:class:`Embed`]
+            A list of embeds to edit the message with.
+        embed: Optional[:class:`Embed`]
+            The embed to edit the message with. ``None`` suppresses the embeds.
+            This should not be mixed with the ``embeds`` parameter.
+        attachments: List[:class:`Attachment`]
+            A list of attachments to keep in the message. If ``[]`` is passed
+            then all attachments are removed.
+        components: Optional[:class:`~discord.ui.MessageComponents`]
+            The updated components to update this message with. If ``None`` is passed then
+            the components are removed.
+
+        Raises
+        -------
+        HTTPException
+            Editing the message failed.
+        TypeError
+            You specified both ``embed`` and ``embeds``.
+        InteractionResponded
+            This interaction has already been responded to before.
+        """
+        if self._responded:
+            raise InteractionResponded(self._parent)
+
+        parent = self._parent
+        msg = parent.message
+        state = parent._state
+        message_id = msg.id if msg else None
+        if parent.type is not InteractionType.component:
+            return
+
+        payload = {}
+        if content is not MISSING:
+            if content is None:
+                payload['content'] = None
+            else:
+                payload['content'] = str(content)
+
+        if embed is not MISSING and embeds is not MISSING:
+            raise TypeError('cannot mix both embed and embeds keyword arguments')
+
+        if embed is not MISSING:
+            if embed is None:
+                embeds = []
+            else:
+                embeds = [embed]
+
+        if embeds is not MISSING:
+            payload['embeds'] = [e.to_dict() for e in embeds]
+
+        if attachments is not MISSING:
+            payload['attachments'] = [a.to_dict() for a in attachments]
+
+        if components is not MISSING:
+            if components is None:
+                payload['components'] = []
+            else:
+                payload['components'] = components.to_dict()
+
+        if allowed_mentions is not MISSING:
+            parent_state = self._parent._state
+            if allowed_mentions:
+                if parent_state.allowed_mentions is not None:
+                    payload['allowed_mentions'] = parent_state.allowed_mentions.merge(allowed_mentions).to_dict()
+                else:
+                    payload['allowed_mentions'] = allowed_mentions.to_dict()
+            elif parent_state.allowed_mentions is not None:
+                payload['allowed_mentions'] = parent_state.allowed_mentions.to_dict()
+
+        data = {
+            "type": InteractionResponseType.message_update.value,
+            "data": payload,
+        }
+        self._aiohttp_response.headers["Content-Type"] = "application/json"
+        await self._aiohttp_response.prepare(self._aiohttp_request)
+        await self._aiohttp_response.write(json.dumps(data).encode())
+        await self._aiohttp_response.write_eof()
 
         self._responded = True